openfl/requirements.txt
absl-py==1.0.0
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
brotlipy==0.7.0
cachetools==5.0.0
certifi==2021.10.8
click==8.0.1
cloudpickle==2.0.0
colorama==0.4.4
commonmark==0.9.1
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
docker==5.0.3
dynaconf==3.1.7
entrypoints==0.4
executing==0.8.3
flatten-json==0.1.13
fonttools==4.34.4
gast==0.3.3
google-auth==2.6.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
grpcio-tools==1.34.1
h5py==2.10.0
importlib-metadata==4.11.3
importlib-resources==5.6.0
ipykernel==6.11.0
ipython==8.2.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.1
joblib==1.1.0
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.2.1
jupyter-core==4.9.2
jupyter-server==1.16.0
jupyterlab==3.3.2
jupyterlab-pygments==0.1.2
jupyterlab-server==2.12.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
mkl-fft==1.3.1
mkl-service==2.4.0
nbclassic==0.3.7
nbclient==0.5.13
nbconvert==6.4.5
nbformat==5.2.0
nest-asyncio==1.5.4
notebook==6.4.10
notebook-shim==0.1.0
numpy==1.18.5
oauthlib==3.2.0
opencv-python==4.6.0.66
openfl==1.3
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pip==22.1.2
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.11.2
Pympler==1.0.1
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzmq==22.3.0
requests-oauthlib==1.3.1
rich==9.1.0
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
Send2Trash==1.8.0
setuptools==61.2.0
sniffio==1.2.0
soupsieve==2.3.1
stack-data==0.2.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminado==0.13.3
testpath==0.6.0
threadpoolctl==3.1.0
torch==1.11.0
torchaudio==0.11.0
torchsummary==1.5.1
torchvision==0.12.0
tornado==6.1
tqdm==4.63.1
traitlets==5.1.1
typing-extensions==3.10.0.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
Werkzeug==2.1.0
wheel==0.37.1
wrapt==1.14.0
zipp==3.7.0
